chris feedback
