+++ lib/rules/jest/no-if.js
@@ -1,15 +1,8 @@
 const {docsUrl} = require('../../utilities');
-
-const TEST_FUNCTION_NAMES = [
-  'it',
-  'xit',
-  'fit',
-  'test',
-  'xtest',
-  'describe',
-  'fdescribe',
-  'xdescribe',
-];
+const {
+  TEST_FUNCTION_NAMES,
+  getTestMethodName,
+} = require('../../utilities/jest');
 
 module.exports = {
   meta: {
@@ -24,24 +17,25 @@ module.exports = {
         'Tests should not contain if statements.',
         'This is usually an indication that you',
         'are attempting to test too much at once',
-        'and may want to consider breaking the if',
-        'statement out into a separate test.',
+        'or not testing what you intend to.',
+        'Consider breaking the if statement out',
+        'into a separate test to resolve this error.',
       ].join(' '),
     },
   },
 
   create(context) {
-    let inCallExpression = false;
+    let callExpressionDepth = 0;
 
     return {
       CallExpression(node) {
-        if (notTestFunction(node) || hasEmptyBody(node)) {
+        if (ignore(node)) {
           return;
         }
-        inCallExpression = true;
+        callExpressionDepth++;
       },
       IfStatement(node) {
-        if (!inCallExpression) {
+        if (callExpressionDepth === 0) {
           return;
         }
 
@@ -50,15 +44,22 @@ module.exports = {
           node,
         });
       },
-      'CallExpression:exit': () => {
-        inCallExpression = false;
+      'CallExpression:exit': (node) => {
+        if (ignore(node)) {
+          return;
+        }
+        callExpressionDepth--;
       },
     };
   },
 };
 
+function ignore(node) {
+  return notTestFunction(node) || hasEmptyBody(node);
+}
+
 function notTestFunction(node) {
-  const method = getMethodName(node);
+  const method = getTestMethodName(node);
   return !matchTestFunctionName(method);
 }
 
@@ -75,19 +76,4 @@ function hasEmptyBody(node) {
     node.arguments[1].body.body &&
     node.arguments[1].body.body.length === 0
   );
-}
-
-function getMethodName({callee}) {
-  switch (callee.type) {
-    case 'CallExpression':
-      return callee.callee.object
-        ? callee.callee.object.name
-        : callee.callee.name;
-    case 'Identifier':
-      return callee.name;
-    case 'MemberExpression':
-      return callee.object.name;
-    default:
-      return false;
-  }
 }
