+++ lib/rules/jest/no-vague-titles.js
@@ -1,5 +1,9 @@
 const commonTags = require('common-tags');
 const {docsUrl} = require('../../utilities');
+const {
+  TEST_FUNCTION_NAMES,
+  getTestMethodName,
+} = require('../../utilities/jest');
 
 const VAGUE_TERMS = {
   should: /should/i,
@@ -10,17 +14,6 @@ const VAGUE_TERMS = {
   every: /every/i,
   descriptive: /descriptive/i,
 };
-
-const TEST_FUNCTION_NAMES = [
-  'it',
-  'xit',
-  'fit',
-  'test',
-  'xtest',
-  'describe',
-  'fdescribe',
-  'xdescribe',
-];
 
 module.exports = {
   meta: {
@@ -57,7 +50,7 @@ module.exports = {
     const allowed = options.allow || [];
 
     function isIgnoredFunctionName(node) {
-      const method = getMethodName(node);
+      const method = getTestMethodName(node);
       return ignored.some((ignoredMethod) => ignoredMethod === method);
     }
 
@@ -74,7 +67,7 @@ module.exports = {
       const containingVagueWords = containsVagueWord(description, allowed);
 
       if (containingVagueWords.length) {
-        const method = getMethodName(node);
+        const method = getTestMethodName(node);
         const vagueWords = commonTags.commaListsAnd`${containingVagueWords}`;
         context.report({
           messageId: 'containsVagueWord',
@@ -96,7 +89,7 @@ module.exports = {
 };
 
 function notTestFunction(node) {
-  const method = getMethodName(node);
+  const method = getTestMethodName(node);
   return !matchTestFunctionName(method);
 }
 
@@ -110,21 +103,6 @@ function hasEmptyDescription({arguments: args}) {
     !args[0] ||
     (args[0].type !== 'Literal' && args[0].type !== 'TemplateLiteral')
   );
-}
-
-function getMethodName({callee}) {
-  switch (callee.type) {
-    case 'CallExpression':
-      return callee.callee.object
-        ? callee.callee.object.name
-        : callee.callee.name;
-    case 'Identifier':
-      return callee.name;
-    case 'MemberExpression':
-      return callee.object.name;
-    default:
-      return false;
-  }
 }
 
 function getDescription({arguments: args}) {
