+++ lib/utilities/index.js
