+++ lib/utilities/jest.js
@@ -1,0 +1,30 @@
+const TEST_FUNCTION_NAMES = [
+  'it',
+  'xit',
+  'fit',
+  'test',
+  'xtest',
+  'describe',
+  'fdescribe',
+  'xdescribe',
+];
+
+function getTestMethodName({callee}) {
+  switch (callee.type) {
+    case 'CallExpression':
+      return callee.callee.object
+        ? callee.callee.object.name
+        : callee.callee.name;
+    case 'Identifier':
+      return callee.name;
+    case 'MemberExpression':
+      return callee.object.name;
+    default:
+      return false;
+  }
+}
+
+module.exports = {
+  TEST_FUNCTION_NAMES,
+  getTestMethodName,
+};
