+++ tests/lib/rules/jest/no-if.js
@@ -225,5 +225,18 @@ ruleTester.run('no-if', rule, {
         },
       ],
     },
+    {
+      code: `it('foo', () => {
+        callExpression()
+        if ('bar') {}
+      })
+      `,
+      parser,
+      errors: [
+        {
+          messageId: 'noIf',
+        },
+      ],
+    },
   ],
 });
